+++ poem/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+
+- Improve `EndpointExt::around`.
+
 # [1.0.19] 2021-11-03
 
 - Add `Request::data`, `Request::set_data`, `Response::data` and `Response::set_data` methods.